Remove behavior & added only the essential code, fix imports, fix styles to have a behavior like a paper-input
